requirements.txt
azure-cognitiveservices-speech
google-generativeai